Introduce four categories, including note.
This allows distinct formatting for lines including "note:", which in turn
makes error lines stand out better.  Lines which don't belong to one of
these categories will use the last detected category.

The formatting of "In file included from" at the beginning of the line is
adjusted to the message text formatting, and the subsequent source location
is formatted as file and line number.